Number Guessing Game v1.1
This is the first major change in the overall look of this project. I've added some features like more rounds, calculating and displaying the score after the game ends, and a lot of details that will create this game more interesting. One of the issues is that the hinting algorithm sometimes makes a mistake and displays inaccurate hints instead of helpful ones so I will get onto that next. I am also planning to start studying creating User Interface (GUI) with Java so I can put it into the code and make the program even easier to use.