Add files via upload
This is a system that is built to handle the most basic procedure when a new employee joins a company. It takes personal information and the branch of the company they are working in. The program generates a personalized email and after that, it also generates a password using the Random() library.